+++ sdk/cognitivelanguage/azure-ai-language-questionanswering/samples/sample_chat.py
@@ -29,7 +29,7 @@ def sample_chit_chat():
 
     endpoint = os.environ["AZURE_QUESTIONANSWERING_ENDPOINT"]
     key = os.environ["AZURE_QUESTIONANSWERING_KEY"]
-    knowledgebase_project = os.environ["AZURE_QUESTIONANSWERING_PROJECT"]
+    knowledge_base_project = os.environ["AZURE_QUESTIONANSWERING_PROJECT"]
 
     client = QuestionAnsweringClient(endpoint, AzureKeyCredential(key))
     with client:
@@ -45,9 +45,9 @@ def sample_chit_chat():
             ),
         )
 
-        output = client.query_knowledgebase(
+        output = client.query_knowledge_base(
             first_question,
-            project_name=knowledgebase_project,
+            project_name=knowledge_base_project,
             deployment_name="test"
         )
         best_candidate = [a for a in output.answers if a.confidence_score > 0.9][0]
@@ -70,9 +70,9 @@ def sample_chit_chat():
             include_unstructured_sources=True
         )
 
-        output = client.query_knowledgebase(
+        output = client.query_knowledge_base(
             followup_question,
-            project_name=knowledgebase_project,
+            project_name=knowledge_base_project,
             deployment_name="test"
         )
         print("Q: {}".format(followup_question.question))