+++ sdk/cognitivelanguage/azure-ai-language-questionanswering/tests/test_query_text.py
@@ -4,6 +4,8 @@
 # Licensed under the MIT License.
 # ------------------------------------
 
+import pytest
+
 from azure.core.exceptions import HttpResponseError, ClientAuthenticationError
 from azure.core.credentials import AzureKeyCredential
 
@@ -120,6 +122,60 @@ class QnATests(QuestionAnsweringTest):
 
         with client:
             output = client.query_text(params)
+            assert len(output.answers) == 3
+            confident_answers = [a for a in output.answers if a.confidence_score > 0.9]
+            assert len(confident_answers) == 2
+            assert confident_answers[0].answer_span.text == "two to four hours"
+
+
+    @GlobalQuestionAnsweringAccountPreparer()
+    def test_query_text_with_str_records(self, qna_account, qna_key):
+        client = QuestionAnsweringClient(qna_account, AzureKeyCredential(qna_key))
+        params = {
+            "question": "How long it takes to charge surface?",
+            "records": [
+                "Power and charging. It takes two to four hours to charge the Surface Pro 4 battery fully from an empty state. " +
+                "It can take longer if you’re using your Surface for power-intensive activities like gaming or video streaming while you’re charging it.",
+                "You can use the USB port on your Surface Pro 4 power supply to charge other devices, like a phone, while your Surface charges. "+
+                "The USB port on the power supply is only for charging, not for data transfer. If you want to use a USB device, plug it into the USB port on your Surface.",
+            ],
+            "language": "en"
+        }
+
+        with client:
+            output = client.query_text(params)
+            assert len(output.answers) == 3
+            confident_answers = [a for a in output.answers if a.confidence_score > 0.9]
+            assert len(confident_answers) == 2
+            assert confident_answers[0].answer_span.text == "two to four hours"
+
+    @GlobalQuestionAnsweringAccountPreparer()
+    def test_query_text_overload(self, qna_account, qna_key):
+        client = QuestionAnsweringClient(qna_account, AzureKeyCredential(qna_key))
+
+        with client:
+            with pytest.raises(TypeError):
+                client.query_text(
+                    question="How long it takes to charge surface?",
+                    records=[
+                        "Power and charging. It takes two to four hours to charge the Surface Pro 4 battery fully from an empty state. " +
+                        "It can take longer if you’re using your Surface for power-intensive activities like gaming or video streaming while you’re charging it.",
+                        {
+                            "text": "You can use the USB port on your Surface Pro 4 power supply to charge other devices, like a phone, while your Surface charges. "+
+                                    "The USB port on the power supply is only for charging, not for data transfer. If you want to use a USB device, plug it into the USB port on your Surface.",
+                            "id": "2"
+                        }
+                    ]
+                )
+            output = client.query_text(
+                question="How long it takes to charge surface?",
+                records=[
+                    "Power and charging. It takes two to four hours to charge the Surface Pro 4 battery fully from an empty state. " +
+                    "It can take longer if you’re using your Surface for power-intensive activities like gaming or video streaming while you’re charging it.",
+                    "You can use the USB port on your Surface Pro 4 power supply to charge other devices, like a phone, while your Surface charges. "+
+                    "The USB port on the power supply is only for charging, not for data transfer. If you want to use a USB device, plug it into the USB port on your Surface.",
+                ]
+            )
             assert len(output.answers) == 3
             confident_answers = [a for a in output.answers if a.confidence_score > 0.9]
             assert len(confident_answers) == 2