+++ sdk/monitor/azure-monitor-query/azure/monitor/query/_models.py
@@ -9,7 +9,7 @@ from enum import Enum
 import uuid
 from typing import Any, Optional, List
 
-from ._helpers import construct_iso8601
+from ._helpers import construct_iso8601, process_row
 from ._generated.models import (
     Column as InternalColumn,
     BatchQueryRequest as InternalLogQueryRequest,
@@ -32,7 +32,7 @@ class LogsQueryResultTable(object):
         # type: (str, List[LogsQueryResultColumn], List[List[str]]) -> None
         self.name = name
         self.columns = columns
-        self.rows = rows
+        self.rows = [process_row(self.columns, row) for row in rows]
 
     @classmethod
     def _from_generated(cls, generated):
@@ -121,8 +121,8 @@ class MetricsResult(object):
     :vartype interval: ~datetime.timedelta
     :ivar namespace: The namespace of the metrics that has been queried.
     :vartype namespace: str
-    :ivar resourceregion: The region of the resource that has been queried for metrics.
-    :vartype resourceregion: str
+    :ivar resource_region: The region of the resource that has been queried for metrics.
+    :vartype resource_region: str
     :ivar metrics: Required. The value of the collection.
     :vartype metrics: list[~monitor_query_client.models.Metric]
     """
@@ -132,7 +132,7 @@ class MetricsResult(object):
         self.timespan = kwargs["timespan"]
         self.interval = kwargs.get("interval", None)
         self.namespace = kwargs.get("namespace", None)
-        self.resourceregion = kwargs.get("resourceregion", None)
+        self.resource_region = kwargs.get("resource_region", None)
         self.metrics = kwargs["metrics"]
 
     @classmethod
@@ -144,7 +144,7 @@ class MetricsResult(object):
             timespan=generated.timespan,
             interval=generated.interval,
             namespace=generated.namespace,
-            resourceregion=generated.resourceregion,
+            resource_region=generated.resourceregion,
             metrics=[Metric._from_generated(m) for m in generated.value] # pylint: disable=protected-access
         )
 
@@ -158,13 +158,10 @@ class LogsBatchQuery(object):
     :param query: The Analytics query. Learn more about the `Analytics query syntax
      <https://azure.microsoft.com/documentation/articles/app-insights-analytics-reference/>`_.
     :type query: str
-    :param ~datetime.timedelta duration: The duration for which to query the data. This can also be accompanied
-     with either start_time or end_time. If start_time or end_time is not provided, the current time is
-     taken as the end time.
-    :keyword datetime start_time: The start time from which to query the data. This should be accompanied
-     with either end_time or duration.
-    :keyword datetime end_time: The end time till which to query the data. This should be accompanied
-     with either start_time or duration.
+    :param timespan: The timespan for which to query the data. This can be a timedelta,
+     a timedelta and a start datetime, or a start datetime/end datetime.
+    :type timespan: ~datetime.timedelta or tuple[~datetime.datetime, ~datetime.timedelta]
+     or tuple[~datetime.datetime, ~datetime.datetime]
     :keyword additional_workspaces: A list of workspaces that are included in the query.
      These can be qualified workspace names, workspace Ids, or Azure resource Ids.
     :paramtype additional_workspaces: list[str]
@@ -180,7 +177,7 @@ class LogsBatchQuery(object):
     :paramtype headers: dict[str, str]
     """
 
-    def __init__(self, query, workspace_id, duration=None, **kwargs): #pylint: disable=super-init-not-called
+    def __init__(self, query, workspace_id, timespan, **kwargs): #pylint: disable=super-init-not-called
         # type: (str, str, Optional[str], Any) -> None
         include_statistics = kwargs.pop("include_statistics", False)
         include_visualization = kwargs.pop("include_visualization", False)
@@ -202,9 +199,7 @@ class LogsBatchQuery(object):
             headers['Prefer'] = prefer
         except TypeError:
             headers = {'Prefer': prefer}
-        start = kwargs.pop('start_time', None)
-        end = kwargs.pop('end_time', None)
-        timespan = construct_iso8601(start, end, duration)
+        timespan = construct_iso8601(timespan)
         additional_workspaces = kwargs.pop("additional_workspaces", None)
         self.id = kwargs.get("request_id", str(uuid.uuid4()))
         self.body = {
@@ -307,8 +302,8 @@ class MetricNamespace(object):
     :paramtype type: str
     :keyword name: The name of the namespace.
     :paramtype name: str
-    :keyword metric_namespace_name: The fully qualified namespace name.
-    :paramtype metric_namespace_name: str
+    :keyword fully_qualified_namespace: The fully qualified namespace name.
+    :paramtype fully_qualified_namespace: str
     """
     def __init__(
         self,
@@ -317,27 +312,27 @@ class MetricNamespace(object):
         self.id = kwargs.get('id', None)
         self.type = kwargs.get('type', None)
         self.name = kwargs.get('name', None)
-        self.metric_namespace_name = kwargs.get('metric_namespace_name', None)
+        self.fully_qualified_namespace = kwargs.get('fully_qualified_namespace', None)
 
     @classmethod
     def _from_generated(cls, generated):
         if not generated:
             return cls()
-        metric_namespace_name = None
+        fully_qualified_namespace = None
         if generated.properties:
-            metric_namespace_name = generated.properties.metric_namespace_name
+            fully_qualified_namespace = generated.properties.metric_namespace_name
         return cls(
             id=generated.id,
             type=generated.type,
             name=generated.name,
-            metric_namespace_name=metric_namespace_name
+            fully_qualified_namespace=fully_qualified_namespace
         )
 
 class MetricDefinition(object):
     """Metric definition class specifies the metadata for a metric.
 
-    :keyword is_dimension_required: Flag to indicate whether the dimension is required.
-    :paramtype is_dimension_required: bool
+    :keyword dimension_required: Flag to indicate whether the dimension is required.
+    :paramtype dimension_required: bool
     :keyword resource_id: the resource identifier of the resource that emitted the metric.
     :paramtype resource_id: str
     :keyword namespace: the namespace the metric belongs to.
@@ -368,7 +363,7 @@ class MetricDefinition(object):
         **kwargs
     ):
         # type: (Any) -> None
-        self.is_dimension_required = kwargs.get('is_dimension_required', None) # type: Optional[bool]
+        self.dimension_required = kwargs.get('dimension_required', None) # type: Optional[bool]
         self.resource_id = kwargs.get('resource_id', None) # type: Optional[str]
         self.namespace = kwargs.get('namespace', None) # type: Optional[str]
         self.name = kwargs.get('name', None) # type: Optional[str]
@@ -387,7 +382,7 @@ class MetricDefinition(object):
         if generated.dimensions is not None:
             dimensions = [d.value for d in generated.dimensions]
         return cls(
-            is_dimension_required=generated.is_dimension_required,
+            dimension_required=generated.is_dimension_required,
             resource_id=generated.resource_id,
             namespace=generated.namespace,
             name=generated.name.value,
@@ -557,9 +552,9 @@ class MetricAvailability(object):
     """Metric availability specifies the time grain (aggregation interval or frequency)
     and the retention period for that time grain.
 
-    :keyword time_grain: the time grain specifies the aggregation interval for the metric. Expressed
+    :keyword granularity: the time grain specifies the aggregation interval for the metric. Expressed
      as a duration 'PT1M', 'P1D', etc.
-    :paramtype time_grain: ~datetime.timedelta
+    :paramtype granularity: ~datetime.timedelta
     :keyword retention: the retention period for the metric at the specified timegrain. Expressed as
      a duration 'PT1M', 'P1D', etc.
     :paramtype retention: ~datetime.timedelta
@@ -569,7 +564,7 @@ class MetricAvailability(object):
         **kwargs
     ):
         # type: (Any) -> None
-        self.time_grain = kwargs.get('time_grain', None)
+        self.granularity = kwargs.get('granularity', None)
         self.retention = kwargs.get('retention', None)
 
     @classmethod
@@ -577,7 +572,7 @@ class MetricAvailability(object):
         if not generated:
             return cls()
         return cls(
-            time_grain=generated.time_grain,
+            granularity=generated.time_grain,
             retention=generated.retention
         )