+++ sdk/monitor/azure-monitor-query/tests/async/test_metrics_client_async.py
@@ -20,8 +20,7 @@ async def test_metrics_auth():
     response = await client.query(
         os.environ['METRICS_RESOURCE_URI'],
         metric_names=["MatchedEventCount"],
-        start_time=datetime(2021, 6, 21),
-        duration=timedelta(days=1),
+        timespan=timedelta(days=1),
         aggregations=[AggregationType.COUNT]
         )
     assert response