+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -9,6 +9,11 @@
 ### Bugs Fixed
 
 ### Other Changes
+- Added context manager methods and `close()` to credentials in the
+  `azure.identity` namespace. At the end of a `with` block, or when `close()`
+  is called, these credentials close their underlying transport sessions.
+  ([#18798](https://github.com/Azure/azure-sdk-for-python/issues/18798))
+
 
 ## 1.7.0b3 (2021-08-10)