+++ sdk/synapse/azure-synapse-managedprivateendpoints/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 0.5.0 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 0.4.0 (2021-08-10)
 
 - Updated API version to "2020-12-01" which is the default API version